Adds application class, rearranges base-android, and introduces the demo package hierarchy
