+++ .idea/gradle.xml
@@ -14,6 +14,9 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/base-android" />
+            <option value="$PROJECT_DIR$/demo" />
+            <option value="$PROJECT_DIR$/demo/logging-android" />
+            <option value="$PROJECT_DIR$/feature/about" />
           </set>
         </option>
       </GradleProjectSettings>
