--- base-android/src/androidTest/java/dev/jjerrell/android/playground/base/android/ExampleInstrumentedTest.kt
@@ -1,23 +1,0 @@
-/* (C) 2023 Jacob Jerrell */
-package dev.jjerrell.android.playground.base.android
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.*
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("dev.jjerrell.android.playground.base.android.test", appContext.packageName)
-    }
-}
